[OPENENGSB-3213] applied refactoring and checked with prePush
