+++ components/api/src/main/java/org/openengsb/core/api/model/ModelDescription.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.openengsb.core.ekb.api;
+package org.openengsb.core.api.model;
 
 import org.osgi.framework.Version;
