+++ components/ekb-api/src/main/java/org/openengsb/core/ekb/api/transformation/TransformationDescription.java
@@ -21,7 +21,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-import org.openengsb.core.ekb.api.ModelDescription;
+import org.openengsb.core.api.model.ModelDescription;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
