+++ components/ekb/src/main/java/org/openengsb/core/ekb/impl/internal/graph/EKBModelGraph.java
@@ -24,7 +24,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
 
-import org.openengsb.core.ekb.api.ModelDescription;
+import org.openengsb.core.api.model.ModelDescription;
 import org.openengsb.core.ekb.api.transformation.TransformationDescription;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
