+++ itests/src/test/java/org/openengsb/itests/exam/ModelTransformationIT.java
@@ -31,6 +31,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.openengsb.core.api.AliveState;
 import org.openengsb.core.api.context.ContextHolder;
+import org.openengsb.core.api.model.ModelDescription;
 import org.openengsb.core.api.model.OpenEngSBModelEntry;
 import org.openengsb.core.workflow.api.RuleManager;
 import org.openengsb.core.workflow.api.WorkflowService;
@@ -39,7 +40,6 @@ import org.openengsb.core.workflow.api.model.RuleBaseElementType;
 import org.openengsb.core.common.AbstractOpenEngSBService;
 import org.openengsb.core.common.transformations.TransformationUtils;
 import org.openengsb.core.common.util.ModelUtils;
-import org.openengsb.core.ekb.api.ModelDescription;
 import org.openengsb.core.ekb.api.TransformationEngine;
 import org.openengsb.core.ekb.api.transformation.TransformationDescription;
 import org.openengsb.domain.example.ExampleDomain;
